ResumeParser/requirements.txt
spacy==3.8.7 
nltk==3.9.1 
pdfplumber==0.11.4
python-docx==1.1.2 
numpy==1.26.4 
docling==2.28.1 
packaging==24.1 
streamlit==1.44.0